+++ libnetwork/sandbox_dns_windows.go
@@ -20,18 +20,8 @@ func (sb *Sandbox) updateHostsFile(ifaceIP []string) error {
 	return nil
 }
 
-func (sb *Sandbox) addHostsEntries(recs []etchosts.Record) {}
-
 func (sb *Sandbox) deleteHostsEntries(recs []etchosts.Record) {}
 
 func (sb *Sandbox) updateDNS(ipv6Enabled bool) error {
-	return nil
-}
-
-func (sb *Sandbox) setupDNS() error {
-	return nil
-}
-
-func (sb *Sandbox) rebuildDNS() error {
 	return nil
 }
